+++ packages/ui/data-entry/input/src/index.ts
@@ -1,0 +1,2 @@
+export { Input, type InputProps } from './Input'
+export { InputPassword, type InputPasswordProps } from './InputPassword'