+++ packages/ui/primitive/collection/__stories__/Collection.stories.tsx
@@ -1,2 +1,10 @@
 import React from 'react'
 import { Collection, type CollectionProps } from '../src'
+
+export default {
+  title: 'Primitive/Collection',
+  component: Collection,
+  tags: ['autodocs']
+}
+
+export const Default = (args: CollectionProps) => <Collection {...args} />