+++ packages/kits/typography/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @negative-space/typography
 
+## 2.0.10
+
+### Patch Changes
+
+- @negative-space/heading@1.2.1
+- @negative-space/text@1.2.1
+
 ## 2.0.9
 
 ### Patch Changes