+++ packages/ui/data-entry/radio/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @negative-space/radio
 
+## 1.5.1
+
+### Patch Changes
+
+- Updated dependencies [b187578]
+  - @negative-space/system@1.3.0
+  - @negative-space/field@1.0.1
+  - @negative-space/collection@2.0.1
+
 ## 1.5.0
 
 ### Minor Changes